Move PHP API to more generic namespace
